pyproject.toml
[project]
name = "pg-relational-transfer"
version = "0.1.0"
description = "Relational data transfer tool for PostgreSQL"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "click>=8.2.1",
    "deepmerge>=2.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
